🐛 fix: fix 0 search results with specific search engine (#6487)
* update

* improve plugin loading state

* improve loading style

* update favicon

* improve search display

* add search results

* fix lint
