+++ src/store/chat/slices/builtinTool/actions/index.ts
@@ -1,0 +1,18 @@
+import { StateCreator } from 'zustand/vanilla';
+
+import { ChatStore } from '@/store/chat/store';
+
+import { ChatDallEAction, dalleSlice } from './dalle';
+import { SearchAction, searchSlice } from './searXNG';
+
+export interface ChatBuiltinToolAction extends ChatDallEAction, SearchAction {}
+
+export const chatToolSlice: StateCreator<
+  ChatStore,
+  [['zustand/devtools', never]],
+  [],
+  ChatBuiltinToolAction
+> = (...params) => ({
+  ...dalleSlice(...params),
+  ...searchSlice(...params),
+});
