+++ src/tools/web-browsing/const.ts
@@ -1,11 +1,12 @@
 export const ENGINE_ICON_MAP: Record<string, string> = {
-  arxiv: 'https://arxiv.org/static/browse/0.3.4/images/icons/favicon-32x32.png',
-  bilibili: 'https://www.bilibili.com/favicon.ico',
-  bing: 'https://www.bing.com/favicon.ico',
-  brave: 'https://brave.com/static-assets/images/brave-favicon.png',
-  duckduckgo: 'https://www.duckduckgo.com/favicon.ico',
-  google: 'https://www.google.com/favicon.ico',
-  npm: 'https://static-production.npmjs.com/da3ab40fb0861d15c83854c29f5f2962.png',
-  qwant: 'https://www.qwant.com/favicon.ico',
-  youtube: 'https://www.youtube.com/favicon.ico',
+  'arxiv': 'https://icons.duckduckgo.com/ip3/arxiv.org.ico',
+  'bilibili': 'https://icons.duckduckgo.com/ip3/bilibili.com.ico',
+  'bing': 'https://icons.duckduckgo.com/ip3/www.bing.com.ico',
+  'brave': 'https://icons.duckduckgo.com/ip3/brave.com.ico',
+  'duckduckgo': 'https://icons.duckduckgo.com/ip3/www.duckduckgo.com.ico',
+  'google': 'https://icons.duckduckgo.com/ip3/google.com.ico',
+  'google scholar': 'https://icons.duckduckgo.com/ip3/scholar.google.com.ico',
+  'npm': 'https://icons.duckduckgo.com/ip3/npmjs.com.ico',
+  'qwant': 'https://icons.duckduckgo.com/ip3/www.qwant.com.ico',
+  'youtube': 'https://icons.duckduckgo.com/ip3/youtube.com.ico',
 };
